pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piradar"
version = "0.1.0"
description = "BGT60TR13C radar sensor driver for Raspberry Pi Dreamhat"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Juergen Hasch", email = "juergen.hasch@mmwaves.de"}
]
maintainers = [
    {name = "Juergen Hasch", email = "juergen.hasch@mmwaves.de"}

]
keywords = ["radar", "bgt60tr13c", "spi", "gpio", "raspberry-pi", "mmw", "fmcw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
requires-python = ">=3.9"
# Base dependencies (client-only, no hardware required)
dependencies = [
    "numpy",
    "numba",
    "pydantic",
    "pyzmq",
    "tqdm",
    "pyyaml",
    "h5py",
    "psutil",
]

[project.optional-dependencies]
# Hardware dependencies (for Raspberry Pi with radar sensor)
hardware = [
    "gpiod>=2.3.0",
    "smbus2>=0.5.0",
    "spidev>=3.6",
]
# Client-only installation (remote access via ZMQ)
client = [
    "matplotlib>=3.5.0",
    "scipy>=1.9.0",
]
# Full installation with all features
all = [
    "gpiod>=2.3.0",
    "smbus2>=0.5.0",
    "spidev>=3.6",
    "matplotlib>=3.5.0",
    "scipy>=1.9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "httpx",  # For API testing
]
docs = [
    "mkdocs>=1.6",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/juhasch/piradar"
Documentation = "https://piradar.readthedocs.io/"
Repository = "https://github.com/juhasch/piradar.git"
"Bug Tracker" = "https://github.com/juhasch/piradar/issues"
"Source Code" = "https://github.com/juhasch/piradar"


[project.scripts]
piradar = "piradar.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["piradar*"]
exclude = ["tests*", "examples*", "MMW-HAT-Release*", "ifx*", "datasheet*"]

[tool.setuptools.package-data]
piradar = ["*.txt", "*.md", "*.yaml", "*.yml"]
"piradar.hw" = ["*.txt", "*.md", "*.yaml", "*.yml"]
"piradar.client" = ["*.txt", "*.md", "*.yaml", "*.yml"]
"piradar.client.config" = ["*.yaml", "*.yml"]

[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "*.pyo", "__pycache__", "*.so", "*.dylib", "*.dll"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "hardware: marks tests that require hardware",
    "asyncio: marks tests as asyncio tests",
]

[tool.coverage.run]
source = ["piradar"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
